requirements.txt
Flask==0.12.2
GoogleAppEngineCloudStorageClient==1.9.22.1